requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
PyYAML~=6.0.1
python-dotenv~=1.0.1
pycryptodome~=3.21.0
ddddocr~=1.5.6
selenium~=4.28.1
PyExecJS2~=1.6.1
urllib3~=2.4.0
gmssl~=3.2.2
httpx~=0.28.1
aiohttp~=3.11.12
certifi~=2025.1.31
pandas~=2.2.3
tenacity~=9.0.0